feat: OrderTablesValidator 구현, Order 패키지 의존성 제거
